Working parsing test case.
To add more parser cases to check, simply add .py files to the testdata/ subdirectory and run
the parser test. The first time around, it will create a .txt file in the same directory, which
for future runs will be treated as the expected result. (Commit the .txt file too)